Simple demo of adding repulsors to Danger Zone.
Turns out, they work just fine. You can't have more than one of them, but it
does exactly what you would expect. Which is utterly and completely game
breaking, as whoever finds it can basically just hug it and survive while the
Zone kills off everyone else. But hey, it DOES work...

To make this actually balanced, it would need to have hitpoints that go down
as it repels the virus, or something like that. And even then, it would be
insanely OP.